functionality to toggle voya is complete, need to clean up code
