+++ src/visualizer/viz.tsx
@@ -1,0 +1,21 @@
+import { MuiThemeProvider } from "@material-ui/core/styles";
+import * as React from "react";
+// import {render} from "react-dom";
+import { theme } from "./components/MUITheme";
+import { GraphQLVoyager } from ".";
+import schema from "../../demo/schema/schema";
+
+export default class Viz extends React.Component {
+  public render() {
+    return (
+      <MuiThemeProvider theme={theme}>
+        <GraphQLVoyager introspection={schema} />
+      </MuiThemeProvider>
+    );
+  }
+}
+
+// render(
+//     <Viz />,
+//   document.getElementById("viz")
+// );
